+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Fixed pod controller logging to use request namespace/name instead of empty pod object fields when pod is not found
 - Fixed a bug where topology constrains with equal required and preferred levels would cause preferred level not to be found.
 - Fixed GPU memory pods Fair Share and Queue Order calculations
+- Interpret negative or zero half-life value as disabled [#818](https://github.com/NVIDIA/KAI-Scheduler/pull/818) [itsomri](https://github.com/itsomri)
 
 ## [v0.12.0] - 2025-12-24